+++ ChunkyMode/Plugin.cs
@@ -25,7 +25,7 @@ namespace HDeMods {
 		public const string PluginGUID = "com." + PluginAuthor + "." + PluginName;
 		public const string PluginAuthor = "HDeDeDe";
 		public const string PluginName = "ChunkyMode";
-		public const string PluginVersion = "0.4.0";
+		public const string PluginVersion = "0.4.1";
 		public static ChunkyModePlugin instance;
 
 		private void Awake() {