+++ CompressToZip/Settings.cs
@@ -37,7 +37,7 @@ internal static class Settings {
                                                "\t\t\"RiskofThunder-R2API_Networking-1.0.2\",\n" +
                                                "\t\t\"RiskofThunder-R2API_Director-2.3.1\",\n" +
                                                "\t\t\"HDeDeDe-HealthComponentAPI-0.2.0\",\n" +
-                                               //"\t\t\"HDeDeDe-InterlopingArtifact-0.1.1\",\n" +
+                                               "\t\t\"HDeDeDe-InterlopingArtifact-0.1.1\",\n" +
                                                "\t\t\"KingEnderBrine-ScrollableLobbyUI-1.9.1\"\n" +
                                                "\t]";
 }